Added external lib for SuperSlot.
The SuperslotSupport project is now integrated as an external library.
This allows us to stop having to deal with a local copy of the support
files.

The DDL has a _D prefix but doesn't mean its a debug DLL, the library
just has the _D name to say that is exports a DLL rather than a static
lib. Might change the name at one point.